+++ src/Application/Adapter/JobQueueInterface.php
@@ -2,9 +2,11 @@
 
 namespace Proximum\Vimeet\Application\Adapter;
 
+use DateTime;
 use Proximum\Vimeet\Domain\Model\Admin;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\File;
+use Proximum\Vimeet\Domain\Model\Happening;
 use Proximum\Vimeet\Domain\Model\Messaging\Campaign;
 use Proximum\Vimeet\Domain\Model\PlannerJob;
 use Proximum\Vimeet\Domain\Model\Sheet;
@@ -222,4 +224,16 @@ interface JobQueueInterface
     public function scheduleUpdateTranslations(?string $emailToNotify = null, ?string $locale = null): void;
 
     public function exportHappeningParticipants(Event $event, Admin $admin, string $locale): void;
+
+    public function zipRecordArchive(
+        Happening $happening,
+        bool $forceRegeneration = false,
+        ?Admin $admin = null,
+        ?string $locale = null
+    ): void;
+
+    public function planDownloadRecordArchive(
+        Happening $happening,
+        DateTime $dueDate
+    ): void;
 }