+++ src/Application/Adapter/VideoConferenceAdapterInterface.php
@@ -44,6 +44,11 @@ interface VideoConferenceAdapterInterface
 
     public function listArchives(string $sessionId): ArchiveList;
 
+    public function listArchiveUrls(string $sessionId): array;
+    public function listArchiveIds(string $sessionId): array;
+
+    public function isRecording(string $sessionId): bool;
+
     /**
      * @param Session            $session
      * @param \DateTimeInterface $endDateTime