+++ src/Application/Event/Events.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Application\Event;
 
 final class Events
@@ -100,6 +92,9 @@ final class Events
     public const HAPPENING_UN_PARTICIPATE = 'happening.un_participate';
     public const HAPPENING_TYPES_UPDATED = 'happening.types.updated';
     public const HAPPENING_DATES_UPDATED = 'happening.dates.updated';
+    public const HAPPENING_ZIP_RECORD_ARCHIVE_NOT_PREPARED = 'happening.webinar.not_zip_record_archive_prepared';
+    public const HAPPENING_ZIP_RECORD_ARCHIVE_PREPARED = 'happening.webinar.zip_record_archive_prepared';
+    public const HAPPENING_RECORDING = 'happening.webinar.recording';
     public const UNAVAILABILITY_ADDED = 'unavailability.added';
     public const UNAVAILABILITY_REMOVED = 'unavailability.removed';
     public const REQUEST_PARTICIPATE = 'request.participate';