+++ src/Domain/Model/Happening.php
@@ -74,10 +74,13 @@ class Happening implements TimeRangeInterface
     private $liveUrl;
 
     /** @var bool */
-    private $sidebarAllowed = true;
+    private $webinarRecorded;
 
     /** @var bool */
-    private $webinarRecorded;
+    private $sidebarAllowed = true;
+
+    /** @var null|string */
+    private $webinarRecordZipFileUrl = null;
 
     public function __construct(
         Event $event,
@@ -112,6 +115,7 @@ class Happening implements TimeRangeInterface
         $this->interactiveWebinar = $interactiveWebinar;
         $this->videoWebinar = $videoWebinar;
         $this->liveUrl = $liveUrl;
+        $this->webinarRecorded = $webinarRecorded;
         $this->sidebarAllowed = $sidebarAllowed;
         $this->webinarRecorded = $webinarRecorded;
     }
@@ -434,5 +438,20 @@ class Happening implements TimeRangeInterface
     public function isVideoWebinarAndHasLiveUrl(): bool
     {
         return $this->isVideoWebinar() && null !== $this->getLiveUrl();
+    }
+
+    public function addWebinarRecordZipFileUrl(?string $url): void
+    {
+        $this->webinarRecordZipFileUrl = $url;
+    }
+
+    public function hasWebinarRecordZipFileUrl(): bool
+    {
+        return !empty($this->getWebinarRecordZipFileUrl());
+    }
+
+    public function getWebinarRecordZipFileUrl(): ?string
+    {
+        return $this->webinarRecordZipFileUrl;
     }
 }