+++ src/Infrastructure/Adapter/VideoConferenceAdapter.php
@@ -13,6 +13,7 @@ use OpenTok\Role;
 use OpenTok\Session;
 use Proximum\Vimeet\Application\Adapter\VideoConferenceAdapterInterface;
 use Proximum\Vimeet\Application\Exception\VideoConference\InvalidTokenGeneratorArgumentsException;
+use Proximum\Vimeet\Domain\Happening\Webinar\RecordStatus;
 
 class VideoConferenceAdapter implements VideoConferenceAdapterInterface
 {
@@ -94,6 +95,48 @@ class VideoConferenceAdapter implements VideoConferenceAdapterInterface
     public function listArchives(string $sessionId): ArchiveList
     {
         return $this->openTok->listArchives(0, null, $sessionId);
+    }
+
+    public function listArchiveUrls(string $sessionId): array
+    {
+        $list = $this->listArchives($sessionId);
+        $urls = [];
+
+        foreach ($list->getItems() as $archive) {
+            $url = $archive->url;
+            if ($url) {
+                $urls[] = $url;
+            }
+        }
+
+        return $urls;
+    }
+
+    public function listArchiveIds(string $sessionId): array
+    {
+        $list = $this->listArchives($sessionId);
+        $archives = $list->getItems();
+
+        usort($archives, static function (Archive $archiveA, Archive $archiveB) {
+            return $archiveA->createdAt <=> $archiveB->createdAt;
+        });
+
+        $ids = [];
+        foreach ($archives as $archive) {
+            $ids[] = $archive->id;
+        }
+
+        return $ids;
+    }
+
+    public function isRecording(string $sessionId): bool
+    {
+        $existingArchives = $this->listArchives($sessionId);
+        $startedArchives = array_filter($existingArchives->getItems(), static function ($archiveItem) {
+            return in_array($archiveItem->status, RecordStatus::IS_RECORDING_STATUS, true);
+        });
+
+        return 0 < count($startedArchives);
     }
 
     /**