+++ src/Infrastructure/Bundle/InfrastructureBundle/Adapter/Happening/Webinar/RecordJobQueueAdapter.php
@@ -2,6 +2,8 @@
 
 namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Adapter\Happening\Webinar;
 
+use DateTime;
+use DateTimeInterface;
 use JMS\JobQueueBundle\Entity\Job;
 use Proximum\Vimeet\Application\Adapter\Happening\Webinar\RecordJobQueueInterface;
 use Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Adapter\AbstractJobQueueAdapter;
@@ -17,7 +19,7 @@ class RecordJobQueueAdapter extends AbstractJobQueueAdapter implements RecordJob
         );
     }
 
-    public function prepareReconciliation(int $happeningId, \DateTimeInterface $reconciliationDate): void
+    public function prepareReconciliation(int $happeningId, DateTimeInterface $reconciliationDate): void
     {
         $command = ReconciliationCommand::NAME;
         $args = [$happeningId];
@@ -25,7 +27,7 @@ class RecordJobQueueAdapter extends AbstractJobQueueAdapter implements RecordJob
         $this->removeJob($command, $args);
 
         $job = new Job($command, $args);
-        $date = new \DateTime();
+        $date = new DateTime();
         $date->setTimestamp($reconciliationDate->getTimestamp());
         $job->setExecuteAfter($date);