+++ src/Ui/Bundle/AdminBundle/Controller/Happening/Webinar/RecordStatusChangeCallbackAction.php
@@ -4,7 +4,6 @@ namespace Proximum\Vimeet\Ui\Bundle\AdminBundle\Controller\Happening\Webinar;
 
 use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
 use Proximum\Vimeet\Application\Command\Happening\Webinar\Record\StatusChangeCallback;
-use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;