+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/VideoConference/Webinar.js
@@ -442,8 +442,16 @@ Webinar.prototype.toggleRecording = function(recording) {
 
     if (recording) {
         this.toggleRecordingButton.classList.add('recording');
+        this.toggleRecordingButton.setAttribute(
+            'title',
+            this.toggleRecordingButton.getAttribute('data-button-recording-title')
+        );
     } else {
         this.toggleRecordingButton.classList.remove('recording');
+        this.toggleRecordingButton.setAttribute(
+            'title',
+            this.toggleRecordingButton.getAttribute('data-button-record-title')
+        );
     }
 };