+++ tests/Ui/Bundle/AdminBundle/Controller/Happening/UpdateActionTest.php
@@ -156,6 +156,7 @@ class UpdateActionTest extends TestCase
         $this->happening->isInteractiveWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->isVideoWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->getLiveUrl()->shouldBeCalled()->willReturn(null);
+        $this->happening->isWebinarRecorded()->shouldBeCalled()->willReturn(true);
         $this->happening->isSidebarAllowed()->willReturn(true);
         $this->happening->isWebinarRecorded()->shouldBeCalled()->willReturn(true);
 
@@ -233,6 +234,7 @@ class UpdateActionTest extends TestCase
         $this->happening->isInteractiveWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->isVideoWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->getLiveUrl()->shouldBeCalled()->willReturn(null);
+        $this->happening->isWebinarRecorded()->shouldBeCalled()->willReturn(true);
         $this->happening->isSidebarAllowed()->willReturn(true);
         $this->happening->isWebinarRecorded()->shouldBeCalled()->willReturn(true);